requirements.txt
# --- Core Framework ---
Django==5.0.6

# --- API Development ---
djangorestframework==3.15.1

# --- Configuration & Environment ---
python-decouple==3.8

# --- Web Server & Security ---
django-cors-headers==4.3.1
gunicorn==22.0.0

# --- Database Connector ---
# For connecting to a PostgreSQL database
dj-database-url==2.1.0
psycopg2-binary==2.9.9

# --- Image Handling ---
# Required for Django's ImageField
pillow==10.4.0

# --- Asynchronous Tasks ---
# For running background jobs
celery==5.3.4
redis==5.0.4
